+++ CHANGES.rst
@@ -9,6 +9,8 @@ Unreleased
     slashes mode is disabled for the rule. :issue:`2467`
 -   Fix router so that partial part matches are not allowed
     i.e. ``/2df`` does not match ``/<int>``. :pr:`2470`
+-   Fix router static part weighting, so that simpler routes are matched
+    before more complex ones. :issue:`2471`
 -   Restore ``ValidationError`` to be importable from
     ``werkzeug.routing``. :issue:`2465`